v1.16.2 — Fix Horizon Redis TypeErrors, remove CMS references
- Cast zremrangebyscore/zrevrangebyscore timestamps to string (phpredis strict_types)
- Guard against Redis object returns from llen/zrevrangebyscore with OPT_PREFIX
- Fix AutoScaler ceil() string→float cast
- Remove CMS AI key references from local.example.php stub

+++ src/Horizon/AutoScaler.php
@@ -152,7 +152,7 @@ class AutoScaler
 
         $totalProcessCount = $pool->totalProcessCount();
 
-        $desiredProcessCount = ceil($workers);
+        $desiredProcessCount = (int) ceil((float) $workers);
 
         if ($desiredProcessCount > $totalProcessCount) {
             $maxUpShift = min(
